Added mounts.py yo unify access points management. Mounts errors and permissions  fixed
