+++ examples/coding/code_agent_biopython.py
@@ -2,17 +2,16 @@ from pathlib import Path
 from dotenv import load_dotenv
 from just_agents.interfaces.IAgent import build_agent, IAgent
 from just_agents.llm_session import LLMSession
+
 from llm_sandbox.micromamba import MicromambaSession
 from llm_sandbox.docker import SandboxDockerSession
 from docker.types import Mount
 import os
 
+from examples.coding.mounts import make_mounts, input_dir, output_dir, coding_examples_dir
 from examples.coding.tools import write_thoughts_and_results
-
 load_dotenv(override=True)
 
-coding_examples_dir = Path(__file__).parent.absolute()
-output_dir = coding_examples_dir / "output"
 
 """
 This example shows how to use a Chain Of Thought code agent to run python code and bash commands.
